Added NoSQL through UnQlite, this lets it be more memory efficient and easier to update in the future.
NoSQL added using UnQlite, which makes it more memory efficient than in-memory, and makes it easier to update in the future. Example files included (sitedb, searchdb). YAML file is for queue storing for web scraper. TODO: need to make the GUI not a GUI anymore, make it so that RPi can run scraper and interrupt via keyboard and/or network activity. Site likely not needed anymore. Should change starting site to Wikipedia (philosophy)